created separate route and controller for blog related functions. updated routes in views to properly work with their new routes
