+++ server.js
@@ -12,7 +12,8 @@ const registerRoutes = require('./routes/registerRoutes'); // import the registe
 const likeRoutes = require('./routes/likesRoutes'); // import the likes routes
 const profileRoutes = require('./routes/profileRoutes'); // import the profile routes
 const settingsRoutes = require('./routes/settingsRoutes');
-const bookmarkRoutes = require('./routes/bookmarkRoutes')
+const bookmarkRoutes = require('./routes/bookmarkRoutes');
+const blogRoutes = require('./routes/blogRoutes')
 app.locals.moment = require('moment'); // allows us to format our date
 
 require('dotenv').config({path: './config/.env'});  // dotenv is used to store the secret keys in a .env file
@@ -46,6 +47,7 @@ app.use(passport.session())
 
 // Routes
 app.use('/', mainRoutes)
+app.use('/blog', blogRoutes)
 app.use('/login', loginRoutes)
 app.use('/register', registerRoutes)
 app.use('/like', likeRoutes)
