+++ views/component/navbar.ejs
@@ -10,6 +10,6 @@
         <li><a href="#"><i class="fa-solid fa-bell fa-lg"></i></a></li>
         <li><a href="/bookmark"><i class="fa-solid fa-bookmark fa-lg"></i></a></li>
         <li class="comment"><a href="#"><i class="fa-solid fa-envelope fa-lg"></i></a></li>
-        <li class="upload"><a href="/new"><i class="fa-solid fa-circle-plus fa-lg"></i></a></li>
+        <li class="upload"><a href="/blog/new"><i class="fa-solid fa-circle-plus fa-lg"></i></a></li>
     </ul>
 </nav>
