--- views/viewport/article.ejs
@@ -1,17 +1,0 @@
-
-<div class="viewContainer">
-    <div class="headerContainer">
-        <h1><%= blog.title %></h1>
-    </div>
-    <div class="bodyContainer">
-        <h4 class="subContainer">by
-            <a href="#">
-                <%= blog.author %>
-            </a> on 
-            <span class="date">
-                <%= blog.createdAt.toLocaleDateString() %>
-            </span>
-        </h4>
-    </div>
-    <%- blog.sanitizedHtml %>
-</div>
