+++ views/viewport/newPost.ejs
@@ -2,7 +2,7 @@
 <link rel="stylesheet" href="/css/blog.css"/>
 <div class="viewContainer">
     <div class="bodyContainer">
-        <form action="/create" enctype="multipart/form-data" method="POST">
+        <form action="/blog/create" enctype="multipart/form-data" method="POST">
             <div class="submitContainer">
                 <input type="submit" value="Publish" class="btn btn-primary mt-2">
             </div>
