+++ views/viewport/postPreview.ejs
@@ -3,13 +3,13 @@
     <% articles.forEach(article => { %>
     <div class="postcardContainer">
         <figure class="postcard">
-            <form action="/<%= article.id %>?_method=DELETE" method="POST">
+            <form action="/blog/<%= article.id %>?_method=DELETE" method="POST">
                 <button type="submit"><i class="fa-solid fa-trash"></i></button>
             </form>
             <%# if (article.author === req.user.username || req.user.role === 'admin') { %>
-            <div><a href="/edit/<%= article.id %>"><i class="fa-solid fa-pencil"></i></a></div>
+            <div><a href="/blog/edit/<%= article.id %>"><i class="fa-solid fa-pencil"></i></a></div>
             <%# } %>
-            <a class="title" href="/<%= article.slug %>/article">
+            <a class="title" href="/blog/<%= article.slug %>/article">
                 <h2><%= article.title %></h2>
                 <h4><%= article.author.userName %></h4>
                 <figcaption>
